server/requirements.txt
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
openai==1.76.0
pydantic==2.11.3